requirements.txt
fastapi[all]
uvicorn[standard]
python-multipart
torch
transformers
sentence-transformers
PyMuPDF
supabase
youtube_search
python-dotenv
